tests: split catch-all test files into one-per-module
The 'one test file per module' convention from tests/README.md was
being violated by three catch-all dump files. Split each one into
proper per-module test files.

Before (3 catch-all files, 12 modules tested):
  tests/test_remaining.py        (webhooks, poster, convert)
  tests/test_more_modules.py     (ergast, omdb, github, request, logs)
  tests/test_api_integrations.py (trakt, anidb, anilist, mal)

After (12 per-module files):
  tests/test_webhooks.py
  tests/test_poster.py
  tests/test_convert.py
  tests/test_ergast.py
  tests/test_omdb.py
  tests/test_github.py
  tests/test_request.py
  tests/test_logs.py
  tests/test_trakt.py
  tests/test_anidb.py
  tests/test_anilist.py
  tests/test_mal.py

Pure 1:1 migration: same 31 tests, same behavior, same fixtures.
Test names were lightly improved during the move (more descriptive
where ambiguous).

tests/test_small_modules.py is intentionally kept as a catch-all per
the README convention — its modules are all genuinely tiny (24-65 LoC
each: stevenlu, gotify, ntfy, notifiarr, icheckmovies, tautulli).

Total suite: 580 tests, all passing in 8.7s.
