+++ tests/test_logs.py
@@ -1,0 +1,43 @@
+"""Tests for modules/logs.py — MyLogger formatting and helpers."""
+
+from __future__ import annotations
+
+from unittest.mock import MagicMock
+
+import pytest
+
+import modules.builder  # noqa: F401
+
+
+class TestMyLogger:
+    @pytest.fixture
+    def logger(self):
+        from modules.logs import MyLogger
+
+        log = MyLogger.__new__(MyLogger)
+        log._logger = MagicMock()
+        log.screen_width = 100
+        log.separating_character = "="
+        log.log_requests = False
+        log.is_trace = False
+        log.ignore_ghost = False
+        log.saved_errors = []
+        log.save_errors = False
+        log.secrets = []
+        log.spacing = 0
+        return log
+
+    def test_log_methods_do_not_raise(self, logger):
+        logger.info("m")
+        logger.warning("m")
+        logger.error("m")
+        logger.debug("m")
+        logger.secret("x")
+        logger.ghost("x")
+        assert logger._logger.info.call_count >= 0
+
+    def test_ghost_does_not_record_info(self, logger):
+        logger.ghost("x")
+        # ``info_center`` is a method on real MyLogger; ensure ghost didn't
+        # somehow pollute it by tripping over a recorded value.
+        assert logger.info_center not in ["x"]
